agile-assignment-master/apps/api/pyproject.toml
[project]
name = "agile-assignment-api"
version = "0.1.0"
description = "FastAPI backend foundation for the agile assignment project."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0,<1.0.0",
  "pydantic-settings>=2.6.0,<3.0.0",
  "uvicorn>=0.32.0,<1.0.0"
]

[dependency-groups]
dev = [
  "httpx>=0.28.0,<1.0.0",
  "pytest>=8.3.0,<9.0.0"
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = [
  "app",
  "tests",
  "README.md",
  ".env.example"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
